Made scale policy configurable via marathon labels

+++ src/main/scala/autoscale/model/MarathonTask.scala
@@ -20,8 +20,8 @@ object TaskStats {
 case class Statistics(cpusLimit:Double,
                       cpusSystemTimeSecs: Double,
                       cpusUserTimeSecs:Double,
-                      memLimitBytes:Float,
-                      memRssBytes:Float,
+                      memLimitBytes: Double,
+                      memRssBytes:Double,
                       timestamp: Double)
 
 object Statistics{
